requirements.txt
altair==4.1.0
appnope==0.1.2
argon2-cffi==21.1.0
astor==0.8.1
attrs==21.2.0
backcall==0.2.0
base58==2.1.0
bleach==4.1.0
blinker==1.4
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
click==7.1.1
cloudpickle==2.0.0
croniter==1.0.15
dask==2021.9.1
debugpy==1.5.0
decorator==5.1.0
defusedxml==0.7.1
distributed==2021.9.1
docker==5.0.3
entrypoints==0.3
fsspec==2021.10.1
future==0.18.2
gitdb==4.0.7
GitPython==3.1.24
google-api-core==1.31.3
google-auth==1.35.0
google-cloud-bigquery==1.28.0
google-cloud-core==1.7.2
google-cloud-secret-manager==2.7.2
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==1.3.3
googleapis-common-protos==1.53.0
gpxcsv==0.2.13
gpxpy==1.4.2
grpc-google-iam-v1==0.12.3
grpcio==1.41.0
HeapDict==1.0.1
idna==2.10
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.2
jsonschema==4.1.0
jupyter-client==7.0.6
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keyring==9.3.1
keyrings.alt==3.2.0
libcst==0.3.21
locket==0.2.1
lxml==4.6.3
MarkupSafe==2.0.1
marshmallow==3.14.0
marshmallow-oneofschema==3.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
msgpack==1.0.2
mypy-extensions==0.4.3
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.4
numpy==1.21.2
packaging==21.0
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
partd==1.2.0
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prefect==0.15.6
prometheus-client==0.11.0
prompt-toolkit==3.0.20
proto-plus==1.19.5
protobuf==3.17.3
psutil==5.8.0
ptyprocess==0.7.0
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydeck==0.7.0
Pygments==2.10.0
pyicloud==0.10.2
pyparsing==2.4.7
pyrsistent==0.18.0
python-box==5.4.1
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.3
pytzdata==2020.1
PyYAML==6.0
pyzmq==22.3.0
requests==2.25.1
rsa==4.7.2
Send2Trash==1.8.0
six==1.16.0
smmap==4.0.0
sortedcontainers==2.4.0
streamlit==1.0.0
tabulate==0.8.9
tblib==1.7.0
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==5.1.0
typing-extensions==3.10.0.2
typing-inspect==0.7.1
tzlocal==2.0.0
urllib3==1.26.7
validators==0.18.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
widgetsnbextension==3.5.1
zict==2.0.0
